requirements.txt
numpy
pandas
scipy
biopython
torch
